fix : "주석 및 불필요한 import삭제, 게시글 생성일 기준 내림차순 및 다른 사용자가 게시글 삭제 할 수 없게 수정"

+++ src/main/java/com/sprta/newsfeed/service/PostServiceImpl.java
@@ -18,6 +18,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -75,7 +76,7 @@ public class PostServiceImpl implements PostService {
     @Override
     //게시글 조회 로직
     public List<PostResponseDto> getAllPosts(int page) {
-        Pageable pageable = PageRequest.of(page, 10);
+        Pageable pageable = PageRequest.of(page, 10, Sort.by("createdAt").descending());
         Page<Post> posts = postRepository.findAll(pageable);
 
         return posts.stream().map(post -> {
@@ -103,12 +104,16 @@ public class PostServiceImpl implements PostService {
         return new PostResponseDto(post, comments, likeCount);
     }
 
-
+    //게시글 삭제
     @Override
-    //게시글 삭제 로직
-    public void deletePost(Long id) {
+    public void deletePost(Long id, String email) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+
+        if (!post.getUser().getEmail().equals(email)) {
+            throw new CustomException(ErrorCode.FORBIDDEN_DELETE);
+        }
+
         postRepository.delete(post);
     }
